fix: care for struct timex uses kernel_long_t
In opposite to the documentation, where `structs timex` is build using
`int` and `long` types, in fact it uses `int` and `__kernel_long_t`
types. This causes for example:

    work/armv7at2hf-neon-oe-linux-gnueabi/pi-hole-ftl/6.1/git/src/ntp/client.c:243:30: error: format '%ld' expects argument of type 'long int', but argument 4 has type 'long long int' [-Werror=format=]
      243 |         log_debug(DEBUG_NTP, "Gradually adjusting system time by %ld us", tx.offset);
          |                              ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  ~~~~~~~~~
          |                                                                             |
          |                                                                             long long int

which can be fixed using PRI formatting from inttypes.h

+++ src/ntp/client.c
@@ -240,7 +240,7 @@ static bool settime_skew(const double offset)
 	tx.offset = 1000000 * offset;
 	tx.modes = ADJ_OFFSET_SINGLESHOT;
 
-	log_debug(DEBUG_NTP, "Gradually adjusting system time by %ld us", tx.offset);
+	log_debug(DEBUG_NTP, "Gradually adjusting system time by %"PRId64" us", (int64_t)tx.offset);
 
 	if(adjtimex(&tx) < 0)
 	{
